refactor: изменил структуру
