+++ src/Products/Detail/Thumbnails/Separator.jsx
