+++ src/Products/List/Filters/Size.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Item from '../Item';
+import Item from './Button';
 
 const renderSizeItems = (items = []) => (
    items.map(item => <Item key={item.name} item={item} />)
