--- src/Sidebar/Logo/styles.css
@@ -1,5 +1,0 @@
-.logo {
-  height: 94px;
-  margin-top: 31px;
-}
-
