--- src/Sidebar/Menu/Item/SubItems/styles.css
@@ -1,17 +1,0 @@
-.menu-sub-items {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 15px 0 0 0;
-}
-
-.menu-sub-items .menu-sub-item {
-  font-family: AndaleMono;
-  font-weight: normal;
-  margin-bottom: 25px;
-}
-
-.menu-sub-items .menu-sub-item:last-child {
-  margin-bottom: 10px;
-}
-
