--- src/Sidebar/Menu/Item/styles.css
@@ -1,29 +1,0 @@
-.menu-item {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 40px;
-}
-
-.menu .menu-item-link a, .menu .menu-sub-item a {
-  color: #303030;
-  text-decoration: none;
-}
-
-.menu .menu-item-link a:hover, .menu .menu-sub-item a:hover {
-  text-decoration: none;
-  color: #ffffff;
-}
-
-.menu .menu-item-link a {
-  font-family: AvenirNext-Bold;
-}
-
-.menu a.menu-item-link_active, .menu .menu-sub-item .menu-sub-item_active {
-  color: #ffffff;
-}
-
-.menu-item-link_active:after {
-  content: url('./images/expanded.png');
-  vertical-align: 10%;
-}
