+++ src/Sidebar/Menu/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Item from './Item';
-import './styles.css';
 
 const renderMenuItems = (items = []) => (
   items.map(item => <Item key={item.id} item={item} />)
