--- src/Sidebar/Menu/styles.css
@@ -1,8 +1,0 @@
-.menu {
-  font-size: 24px;
-  margin-top: 150px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
