+++ src/Sidebar/styles.css
@@ -8,3 +8,97 @@
   align-items: center;
 }
 
+.search-form {
+  flex-basis: 100%;
+  display: flex;
+}
+
+.search-submit {
+  width: 28px;
+  height: 30px;
+  background-color: transparent;
+  border: none;
+  background-image: url('./images/search_submit.png');
+  color: transparent;
+  cursor: pointer;
+  padding-right: 5px;
+}
+
+.search-input-container {
+  padding-left: 15px;
+  flex-basis: 100%;
+}
+
+.search-input {
+  background-color: transparent;
+  border: none;
+  color: white;
+  height: 30px;
+  font-size: 24px;
+}
+
+.search-input:focus {
+  outline: none;
+}
+
+.menu {
+  font-size: 24px;
+  margin-top: 150px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.logo {
+  height: 94px;
+  margin-top: 31px;
+}
+
+.menu-sub-items {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 15px 0 0 0;
+}
+
+.menu-sub-items .menu-sub-item {
+  font-family: AndaleMono;
+  font-weight: normal;
+  margin-bottom: 25px;
+}
+
+.menu-sub-items .menu-sub-item:last-child {
+  margin-bottom: 10px;
+}
+
+.menu-item {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 40px;
+}
+
+.menu .menu-item-link a, .menu .menu-sub-item a {
+  color: #303030;
+  text-decoration: none;
+}
+
+.menu .menu-item-link a:hover, .menu .menu-sub-item a:hover {
+  text-decoration: none;
+  color: #ffffff;
+}
+
+.menu .menu-item-link a {
+  font-family: AvenirNext-Bold;
+}
+
+.menu a.menu-item-link_active, .menu .menu-sub-item .menu-sub-item_active {
+  color: #ffffff;
+}
+
+.menu-item-link_active:after {
+  content: url('./images/expanded.png');
+  vertical-align: 10%;
+}
+
+
