+++ src/components/SaleBadge.jsx
